Move JS modules to subdir
