+++ pronto/static/modules/method.js
@@ -1,4 +1,4 @@
-import * as utils from './utils.js';
+import * as utils from '../utils.js';
 
 function calcPixelSlope(p1, p2, distance) {
     return {
