+++ pronto/templates/comments.html
@@ -27,7 +27,7 @@
 
 <script src="/static/vendor/jquery.min.js"></script>
 <script src="/static/vendor/semantic.min.js"></script>
-<script src="/static/comments.js" type="module"></script>
+<script src="/static/modules/comments.js" type="module"></script>
 
 </body>
 </html>
