+++ pronto/templates/database.html
@@ -157,6 +157,6 @@
 
 <script src="/static/vendor/jquery.min.js"></script>
 <script src="/static/vendor/semantic.min.js"></script>
-<script src="/static/database.js" type="module"></script>
+<script src="/static/modules/database.js" type="module"></script>
 </body>
 </html>
