+++ pronto/templates/database2.html
@@ -78,6 +78,6 @@
 
 <script src="/static/vendor/jquery.min.js"></script>
 <script src="/static/vendor/semantic.min.js"></script>
-<script src="/static/database2.js" type="module"></script>
+<script src="/static/modules/database2.js" type="module"></script>
 </body>
 </html>
