+++ pronto/templates/descriptions.html
@@ -44,7 +44,7 @@
 
 <script src="/static/vendor/jquery.min.js"></script>
 <script src="/static/vendor/semantic.min.js"></script>
-<script src="/static/descriptions.js" type="module"></script>
+<script src="/static/modules/descriptions.js" type="module"></script>
 
 </body>
 </html>
