+++ pronto/templates/entry.html
@@ -435,7 +435,7 @@
 
 <script src="/static/vendor/jquery.min.js"></script>
 <script src="/static/vendor/semantic.min.js"></script>
-<script src="/static/entry.js" type="module"></script>
+<script src="/static/modules/entry.js" type="module"></script>
 
 </body>
 </html>
