+++ pronto/templates/enzymes.html
@@ -44,7 +44,7 @@
 
 <script src="/static/vendor/jquery.min.js"></script>
 <script src="/static/vendor/semantic.min.js"></script>
-<script src="/static/enzymes.js" type="module"></script>
+<script src="/static/modules/enzymes.js" type="module"></script>
 
 </body>
 </html>
