+++ pronto/templates/go.html
@@ -65,7 +65,7 @@
 
 <script src="/static/vendor/jquery.min.js"></script>
 <script src="/static/vendor/semantic.min.js"></script>
-<script src="/static/go.js" type="module"></script>
+<script src="/static/modules/go.js" type="module"></script>
 
 </body>
 </html>
