+++ pronto/templates/index.html
@@ -48,7 +48,7 @@
 
 <script src="/static/vendor/jquery.min.js"></script>
 <script src="/static/vendor/semantic.min.js"></script>
-<script src="/static/index.js" type="module"></script>
+<script src="/static/modules/index.js" type="module"></script>
 
 </body>
 </html>
