+++ pronto/templates/matches.html
@@ -64,7 +64,7 @@
 
 <script src="/static/vendor/jquery.min.js"></script>
 <script src="/static/vendor/semantic.min.js"></script>
-<script src="/static/matches.js" type="module"></script>
+<script src="/static/modules/matches.js" type="module"></script>
 
 </body>
 </html>
