+++ pronto/templates/matrices.html
@@ -38,7 +38,7 @@
 
 <script src="/static/vendor/jquery.min.js"></script>
 <script src="/static/vendor/semantic.min.js"></script>
-<script src="/static/matrices.js" type="module"></script>
+<script src="/static/modules/matrices.js" type="module"></script>
 
 </body>
 </html>
