+++ pronto/templates/method.html
@@ -119,7 +119,7 @@
 
 <script src="/static/vendor/jquery.min.js"></script>
 <script src="/static/vendor/semantic.min.js"></script>
-<script src="/static/method.js" type="module"></script>
+<script src="/static/modules/method.js" type="module"></script>
 
 </body>
 </html>
