+++ pronto/templates/protein.html
@@ -138,7 +138,7 @@
 
 <script src="/static/vendor/jquery.min.js"></script>
 <script src="/static/vendor/semantic.min.js"></script>
-<script src="/static/protein.js" type="module"></script>
+<script src="/static/modules/protein.js" type="module"></script>
 <script>
     const entries = {{ protein.entries|tojson }};
     const methods = {{ protein.methods|tojson }};
