+++ pronto/templates/taxonomy.html
@@ -37,7 +37,7 @@
 
 <script src="/static/vendor/jquery.min.js"></script>
 <script src="/static/vendor/semantic.min.js"></script>
-<script src="/static/taxonomy.js" type="module"></script>
+<script src="/static/modules/taxonomy.js" type="module"></script>
 
 </body>
 </html>
